requirements.txt
pdf2image>=1.16.0
PyPDF2>=1.26.0
rich>=10.15.2
Pillow>=8.4.0
numpy>=1.21.4
opencv-python>=4.5.4.60
